Projects/crypto_trend_backtester/requirements.txt
numpy
pandas
pyyaml
tqdm